React code: Fixed AudioPlayer.jsx bug.. Removed pause option
Summary:
Lets say we have 2 clips a,b. Play clip-a first and pause it sometime and play the second clip and play clip-b, pause-b and play clip-a. Now a should be played but b gets played
. This bug is happening because of the Sound object which overwritten.
I've just declared it inside the audio player component and two more
props to save the state when the clip is paused.

Test Plan:
1. Inserted a few debug statements and tweaked the AudioPlayer.jsx code

2. Understood the error is due to overwriting of Sound object.

3. Resolved it by declaring it inside.

Reviewers: abhilashi, vamshi

Differential Revision: http://104.155.197.196/D150